两个模型都验证可推理，修四个只有真跑才暴露的问题
新机器上从零部署，两个底座下完（ether0 87.8GB + Qwen 51.8GB），
自检九项全绿，从 Mac 端到端调通：
  成膜添加剂  VC 比 EC 低 0.372 eV、FEC 低 0.566 eV（文献 0.3~0.9）
  条件生成    目标 gap 8.5 → 8.494 / 8.512 / 8.523，偏差全在 ±0.03
  知识层      答对 H2O

修的四个
  ① ether0 的 config.json 写 torch_dtype: float32，不显式指定就按 fp32
     加载，24B 要 88GB，85GB 的卡直接 OOM。改为显式 bf16。
  ② 自检把 CUDA OOM 当「可选项跳过」，照样报「全部通过」——
     用户会拿到自检全绿但两个模型都用不了的环境。
     改为区分「没装」和「装了但用不了」。
  ③ 知识层断言写的是 len(a) < 200，模型答对了但带解释就被判「回答异常」。
     改为判内容（答案里有没有 H2O）。
  ④ 底座完整性判据第三版：改用仓库自己的 index.json 分片清单。
     前两版分别栽在「只看 config.json」和「算了 .cache 的未完成分片，
     且我猜的模型大小本身就错（标 48GB 实际 75GB）」。

硬约束：生成器 bf16 约 44GB + 知识层 54GB = 98GB > 单卡 85GB，
两个模型不能同时驻留，均为懒加载谁用谁载。

+++ clearchem/clearchem.py
@@ -122,8 +122,13 @@ class ClearChem:
         if tok.pad_token is None:
             tok.pad_token = tok.eos_token
         tok.padding_side = "left"
-        base = AutoModelForCausalLM.from_pretrained(SNAP, device_map={"": 0},
-                                                    trust_remote_code=True)
+        # 必须显式指定 bf16：ether0 的 config.json 里 torch_dtype 写的是 float32，
+        # 不指定就按 fp32 加载，24B 要 88 GB —— 85 GB 的卡直接 OOM，
+        # 而且和知识层（27B bf16，54 GB）根本不可能同时驻留。
+        # bf16 之后生成器约 44 GB，两个模型仍不能同时在一张卡上，
+        # 所以 knowledge.py 和这里都做懒加载，谁用谁载。
+        base = AutoModelForCausalLM.from_pretrained(
+            SNAP, device_map={"": 0}, dtype=torch.bfloat16, trust_remote_code=True)
         model = PeftModel.from_pretrained(base, ADP); model.eval()
         hid = model.get_input_embeddings().weight.shape[1]
